Added `toInt` to all values of the uses commands
#18 Notification when consuming flask without uses

+++ src/main/kotlin/de/lordtkay/hyflask/uses/command/GetUsesCommand.kt
@@ -30,8 +30,8 @@ class GetUsesCommand(
 
             is EntityStatUtility.Result.Success ->
                 Message.translation("$translationKey.success")
-                    .param("uses", result.current)
-                    .param("max", result.max)
+                    .param("uses", result.current.toInt())
+                    .param("max", result.max.toInt())
         }
 
         commandContext.sendMessage(message)
