+++ src/main/kotlin/de/lordtkay/hyflask/uses/command/SetMaxUsesCommand.kt
@@ -44,8 +44,8 @@ class SetMaxUsesCommand(
 
             is EntityStatUtility.Result.Success -> {
                 Message.translation("$translationKey.success")
-                    .param("uses", result.current)
-                    .param("max", result.max)
+                    .param("uses", result.current.toInt())
+                    .param("max", result.max.toInt())
             }
         }
